requirements.txt
Flask==3.1.0
marshmallow==3.23.2
requests==2.32.3
gunicorn==23.0.0
PyYAML==6.0.2
python-dotenv==1.0.1
flask-smorest==0.45.0